VideoUpscaler/requirements.txt
boto3
pySqsListener
pillow
python-dotenv